requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
slackclient==2.9.4
yarl==1.7.2
